requirements.txt
# Core Framework
gradio==4.12.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ML/AI
torch==2.2.0
transformers==4.37.0
peft==0.8.0
bitsandbytes==0.42.0
accelerate==0.26.0
sentence-transformers==2.3.1
einops==0.7.0

# LLM Local
ollama==0.1.6

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
markdown==3.5.1
langchain==0.1.5
langchain-community==0.0.19
beautifulsoup4==4.12.3
lxml==5.1.0

# Vector Store & Embeddings
chromadb==0.4.22
faiss-cpu==1.7.4

# Data & Validation
datasets==2.16.1
jsonschema==4.20.0
pandas==2.1.4
numpy==1.26.3

# Monitoring & Logging
tensorboard==2.15.1
wandb==0.16.2
gpustat==1.1.1
nvidia-ml-py==12.535.133
plotly==5.18.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
aiohttp==3.9.1
aiofiles==23.2.1
tqdm==4.66.1
colorama==0.4.6
rich==13.7.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==24.1.0
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
